backend/requirements.txt
# Lawyer Sites - Backend
# Python 3.9+

fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator>=2.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Opcional para Postgres (descomentar si se migra)
# psycopg2-binary==2.9.9
